+++ src/wmseed.h
@@ -18,11 +18,12 @@ typedef struct {
   MiniSeedRecord record[1];
   int data_pending;
   FILE *output;
+  FILE *logfile;
   Time last_t;
   int64_t last_sn;
 } WMSeed;
 
-WMSeed *wmseed_new(const char *template, const char *station, const char *location, const char *channel, const char *network, double sample_rate, int64_t cut);
+WMSeed *wmseed_new(FILE *logfile, const char *template, const char *station, const char *location, const char *channel, const char *network, double sample_rate, int64_t cut);
 int wmseed_destroy(WMSeed *w);
 int wmseed_sample(WMSeed *w, int32_t sample);
 int wmseed_time(WMSeed *w, Time t, int64_t sample_number);
@@ -41,14 +42,29 @@ int wmseed_time(WMSeed *w, Time t, int64_t sample_number);
 #include <unistd.h>
 #include <math.h>
 
-static void *wmseed__allocate(void *x, size_t size)
+#if __GNUC__ || __clang__
+__attribute__((format(printf, 3, 4)))
+#endif
+static void wmseed__log(WMSeed *w, FILE *f, const char *format, ...)
+{
+  va_list args;
+  va_start(args, format);
+  if (w && w->logfile) {
+    vfprintf(w->logfile, format, args);
+    fflush(w->logfile);
+  }
+  vfprintf(f, format, args);
+  fflush(f);
+  va_end(args);
+}
+
+static void *wmseed__allocate(WMSeed *w, void *x, size_t size)
 {
   void *t;
   if (size) {
     t = realloc(x, size);
     if (!t) {
-      fprintf(stderr, "Out of memory!\n");
-      fflush(stderr);
+      wmseed__log(w, stderr, "Out of memory!\n");
       exit(1);
     }
     return t;
@@ -58,11 +74,11 @@ static void *wmseed__allocate(void *x, size_t size)
   }
 }
 
-static char *wmseed__strdup(const char *s)
+static char *wmseed__strdup(WMSeed *w, const char *s)
 {
   char *s2;
   if (s) {
-    s2 = wmseed__allocate(0, strlen(s) + 1);
+    s2 = wmseed__allocate(w, 0, strlen(s) + 1);
     strcpy(s2, s);
     return s2;
   } else {
@@ -71,16 +87,16 @@ static char *wmseed__strdup(const char *s)
 }
 
 #if __GNUC__ || __clang__
-__attribute__((format(printf, 2, 3)))
+__attribute__((format(printf, 3, 4)))
 #endif
-static char *wmseed__strappend(char *s, const char *format, ...)
+static char *wmseed__strappend(WMSeed *w, char *s, const char *format, ...)
 {
   size_t l1 = s ? strlen(s) : 0, l2;
   va_list args;
   va_start(args, format);
   l2 = vsnprintf(0, 0, format, args);
   va_end(args);
-  s = wmseed__allocate(s, l1 + l2 + 1);
+  s = wmseed__allocate(w, s, l1 + l2 + 1);
   va_start(args, format);
   vsnprintf(s + l1, l2 + 1, format, args);
   va_end(args);
@@ -98,54 +114,54 @@ static char *wmseed__filename(WMSeed *w, Time t)
       ++tmpl;
       switch (*tmpl) {
       case '%':
-        s = wmseed__strappend(s, "%%");
+        s = wmseed__strappend(w, s, "%%");
         break;
       case 'y':
-        s = wmseed__strappend(s, "%04d", d.year);
+        s = wmseed__strappend(w, s, "%04d", d.year);
         break;
       case 'm':
-        s = wmseed__strappend(s, "%02d", d.month);
+        s = wmseed__strappend(w, s, "%02d", d.month);
         break;
       case 'd':
-        s = wmseed__strappend(s, "%02d", d.day);
+        s = wmseed__strappend(w, s, "%02d", d.day);
         break;
       case 'h':
-        s = wmseed__strappend(s, "%02d", d.hour);
+        s = wmseed__strappend(w, s, "%02d", d.hour);
         break;
       case 'i':
-        s = wmseed__strappend(s, "%02d", d.min);
+        s = wmseed__strappend(w, s, "%02d", d.min);
         break;
       case 's':
-        s = wmseed__strappend(s, "%02d", d.sec);
+        s = wmseed__strappend(w, s, "%02d", d.sec);
         break;
       case 'j':
-        s = wmseed__strappend(s, "%03d", yday);
+        s = wmseed__strappend(w, s, "%03d", yday);
         break;
       case 'S':
-        s = wmseed__strappend(s, "%s", w->station);
+        s = wmseed__strappend(w, s, "%s", w->station);
         break;
       case 'L':
-        s = wmseed__strappend(s, "%s", w->location);
+        s = wmseed__strappend(w, s, "%s", w->location);
         break;
       case 'C':
-        s = wmseed__strappend(s, "%s", w->channel);
+        s = wmseed__strappend(w, s, "%s", w->channel);
         break;
       case 'N':
-        s = wmseed__strappend(s, "%s", w->network);
+        s = wmseed__strappend(w, s, "%s", w->network);
         break;
       default:
         free(s);
         return 0;
       }
     } else {
-      s = wmseed__strappend(s, "%c", *tmpl);
+      s = wmseed__strappend(w, s, "%c", *tmpl);
     }
     ++tmpl;
   }
   return s;
 }
 
-static char *wmseed__dirname(const char *path)
+static char *wmseed__dirname(WMSeed *w, const char *path)
 {
   size_t l = 0, x = -1;
   char *s;
@@ -154,21 +170,21 @@ static char *wmseed__dirname(const char *path)
     ++l;
   }
   if (x > 1) {
-    s = wmseed__allocate(0, x + 1);
+    s = wmseed__allocate(w, 0, x + 1);
     memcpy(s, path, x);
     s[x] = 0;
   } else {
-    s = wmseed__allocate(0, 2);
+    s = wmseed__allocate(w, 0, 2);
     s[0] = x ? '.' : '/';
     s[1] = 0;
   }
   return s;
 }
 
-static int wmseed__mkdir_p(const char *path)
+static int wmseed__mkdir_p(WMSeed *w, const char *path)
 {
   struct stat sb;
-  char *p = wmseed__strdup(path);
+  char *p = wmseed__strdup(w, path);
   size_t i, len = strlen(p);
 
   // Remove trailing slash.
@@ -204,16 +220,17 @@ static int wmseed__mkdir_p(const char *path)
   return 0;
 }
 
-WMSeed *wmseed_new(const char *template, const char *station, const char *location, const char *channel, const char *network, double sample_rate, int64_t cut)
+WMSeed *wmseed_new(FILE *logfile, const char *template, const char *station, const char *location, const char *channel, const char *network, double sample_rate, int64_t cut)
 {
   WMSeed *w;
-  w = wmseed__allocate(0, sizeof(*w));
+  w = wmseed__allocate(0, 0, sizeof(*w));
+  w->logfile = logfile;
   w->cut = cut * 1000000;
-  w->template = wmseed__strdup(template);
-  w->station = wmseed__strdup(station);
-  w->location = wmseed__strdup(location);
-  w->channel = wmseed__strdup(channel);
-  w->network = wmseed__strdup(network);
+  w->template = wmseed__strdup(w, template);
+  w->station = wmseed__strdup(w, station);
+  w->location = wmseed__strdup(w, location);
+  w->channel = wmseed__strdup(w, channel);
+  w->network = wmseed__strdup(w, network);
   w->sample_rate = sample_rate;
   w->sb = samplebuffer_new();
   w->record_number = 0;
@@ -254,8 +271,7 @@ static void wmseed__flush(WMSeed *w)
 {
   if (w->data_pending) {
     if (fwrite(w->record->data, sizeof(w->record->data), 1, w->output) != 1) {
-      fprintf(stderr, "I/O error!\n");
-      fflush(stderr);
+      wmseed__log(w, stderr, "I/O error!\n");
       exit(1);
     }
     w->data_pending = 0;
@@ -290,19 +306,17 @@ static void wmseed__create_file(WMSeed *w, Time t)
     fclose(w->output);
   }
   filename = wmseed__filename(w, t);
-  dirname = wmseed__dirname(filename);
-  wmseed__mkdir_p(dirname);
+  dirname = wmseed__dirname(w, filename);
+  wmseed__mkdir_p(w, dirname);
   free(dirname);
   w->output = fopen(filename, "wb");
   if (!w->output) {
-    fprintf(stderr, "Could not create file '%s': %s.\n", filename, strerror(errno));
-    fflush(stderr);
+    wmseed__log(w, stderr, "Could not create file '%s': %s.\n", filename, strerror(errno));
     exit(1);
   } else {
-    fprintf(stderr, "Created file '%s'.\n", filename);
+    wmseed__log(w, stderr, "Created file '%s'.\n", filename);
   }
   free(filename);
-
 }
 
 // b must be positive.